Okay lots happening in this commit
I made a VideoGrid, felt the need to made a VideoGridContainer as well
I made a Post component for individual videos in the grid
I changed the initial state in Context just to have more videos to map
I brought a random image into local project files to be a stand in

The biggest challenge was the styling, for sure. I honestly have very
little to say about it. Lots of looking at examples, trial and error,
taking brain-breaks.

The hover effect was a little tricky at first, but I have fallen in love
with programmatically changing styles inside styled components. The real
new thing was realizing I needed to change transform: translateY(25%) to
translateY(0).